requirements.txt
psycopg2-binary>=2.9
requests>=2.28
PyYAML>=6.0
flask>=3.0
